requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# Search & Data Collection (핵심!)
tavily-python>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Financial Data
finance-datareader>=0.9.50
pandas>=2.0.0
numpy>=1.24.0

# Document Processing
python-docx>=1.1.0
pdfminer.six>=20221105

# Visualization
matplotlib>=3.7.0

# Optional (Development)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
